# apps/web/frontend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfiguredIngestionProcessResponseModel": {
        "additionalProperties": false,
        "properties": {
          "promotions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PromotionResultModel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promotions"
          },
          "result": {
            "$ref": "#/components/schemas/ConfiguredIngestionProcessResultModel"
          }
        },
        "required": [
          "result"
        ],
        "title": "ConfiguredIngestionProcessResponseModel",
        "type": "object"
      },
      "ConfiguredIngestionProcessResultModel": {
        "additionalProperties": false,
        "properties": {
          "discovered_files": {
            "title": "Discovered Files",
            "type": "integer"
          },
          "ingestion_definition_id": {
            "title": "Ingestion Definition Id",
            "type": "string"
          },
          "processed_files": {
            "title": "Processed Files",
            "type": "integer"
          },
          "rejected_files": {
            "title": "Rejected Files",
            "type": "integer"
          },
          "run_ids": {
            "items": {
              "type": "string"
            },
            "title": "Run Ids",
            "type": "array"
          }
        },
        "required": [
          "ingestion_definition_id",
          "discovered_files",
          "processed_files",
          "rejected_files",
          "run_ids"
        ],
        "title": "ConfiguredIngestionProcessResultModel",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HaIngestRequest": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "source_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source System"
          },
          "states": {
            "items": {
              "$ref": "#/components/schemas/HaStateObject"
            },
            "title": "States",
            "type": "array"
          }
        },
        "required": [
          "states"
        ],
        "title": "HaIngestRequest",
        "type": "object"
      },
      "HaStateObject": {
        "properties": {
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "last_changed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Changed"
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "entity_id",
          "state"
        ],
        "title": "HaStateObject",
        "type": "object"
      },
...
      "PromotionResultModel": {
        "additionalProperties": false,
        "properties": {
          "facts_loaded": {
            "title": "Facts Loaded",
            "type": "integer"
          },
          "marts_refreshed": {
            "items": {
              "type": "string"
            },
            "title": "Marts Refreshed",
            "type": "array"
          },
          "publication_keys": {
            "items": {
              "type": "string"
            },
            "title": "Publication Keys",
            "type": "array"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "skip_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Reason"
          },
          "skipped": {
            "title": "Skipped",
            "type": "boolean"
          }
        },
        "required": [
          "facts_loaded",
          "marts_refreshed",
          "publication_keys",
          "skipped"
        ],
        "title": "PromotionResultModel",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Homelab Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ha/ingest": {
      "post": {
        "operationId": "ingest_ha_states_api_ha_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HaIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ingest Ha States Api Ha Ingest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Ha States"
      }
    },
...
    "/ingest/ingestion-definitions/{ingestion_definition_id}/process": {
      "post": {
        "operationId": "process_ingestion_definition_ingest_ingestion_definitions__ingestion_definition_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "ingestion_definition_id",
            "required": true,
            "schema": {
              "title": "Ingestion Definition Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguredIngestionProcessResponseModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process Ingestion Definition"
      }
    },
...
    "/landing/{extension_key}": {
      "post": {
        "operationId": "run_landing_extension_landing__extension_key__post",
        "parameters": [
          {
            "in": "path",
            "name": "extension_key",
            "required": true,
            "schema": {
              "title": "Extension Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Landing Extension Landing  Extension Key  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Landing Extension"
      }
    },
...
    "/transformations/{extension_key}": {
      "get": {
        "operationId": "run_transformation_extension_transformations__extension_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "extension_key",
            "required": true,
            "schema": {
              "title": "Extension Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Transformation Extension Transformations  Extension Key  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Transformation Extension"
      }
    }
  }
}